refactor: remove variables
