+++ src/constants/index.ts
@@ -3,3 +3,4 @@ export const ERROR_STATUS = 'This username is already in the game, check that th
 export const ERROR_LENGTH = 'Minimum 5 characters';
 export const ERROR_NAME = 'This username is reserved, please select a different username';
 export const BOT_NAME = 'DEFAULT_PLAYER';
+export const MAX_AMOUNT_PLAYERS_IN_ROOM = 2;
