+++ src/handlers/addShips.ts
@@ -1,12 +1,11 @@
 import { PositionShipsObject } from '../types/interface/position.js';
 import { AddShips } from '../types/interface/addShips.js';
-import { Socket } from '../types/types/common.js';
+import { Socket } from '../types/types/index.js';
 import { TypeData } from '../types/enum/typeData.js';
 import { socketBase, gameRoomsBase } from '../store/index.js';
 import { updateDataShips } from '../helpers/updateDataShips.js';
 import { getResponseObject } from '../helpers/createrObjects.js';
-
-const MAX_AMOUNT_PLAYERS_IN_ROOM = 2;
+import { MAX_AMOUNT_PLAYERS_IN_ROOM } from '../constants/index.js';
 
 export const addShips = (object: AddShips, _socket: Socket) => {
   const objectData = JSON.parse(object.data) as PositionShipsObject;
