+++ src/handlers/attackShips.ts
@@ -1,5 +1,5 @@
 import { AttackClient, AttackData } from '../types/interface/attack.js';
-import { Socket } from '../types/types/common.js';
+import { Socket } from '../types/types/index.js';
 import { checkMovePlayer, checkShootPlace, getDataPlaceShoot } from '../helpers/supportAttack.js';
 import { shoot } from '../helpers/shoot.js';
 import { gameRoomsBase } from '../store/index.js';
