+++ src/handlers/registration.ts
@@ -1,4 +1,4 @@
-import { Socket } from '../types/types/common.js';
+import { Socket } from '../types/types/index.js';
 import { DataReg, RegObject } from '../types/interface/reg.js';
 import { TypeData } from '../types/enum/typeData.js';
 import { getResponseObject } from '../helpers/createrObjects.js';
